feat: Use feature module for rarities with oracle
