+++ .ci/.env.prd
@@ -41,5 +41,4 @@ REACT_APP_FF_ENS=1
 REACT_APP_FF_WEARABLES=1
 REACT_APP_FF_WEARABLES_PUBLISH=1
 REACT_APP_FF_SKINS=1
-REACT_APP_FF_RARITIES_WITH_ORACLE=0
 REACT_APP_UNDER_MAINTAINANCE=
