+++ .env.example
@@ -59,4 +59,3 @@ REACT_APP_FF_ENS=1
 REACT_APP_FF_WEARABLES=1
 REACT_APP_FF_WEARABLES_PUBLISH=1
 REACT_APP_FF_SKINS=1
-REACT_APP_FF_RARITIES_WITH_ORACLE=1
