+++ now.json
@@ -37,7 +37,6 @@
       "REACT_APP_FF_WEARABLES_PUBLISH": "1",
       "REACT_APP_FF_ENS": "1",
       "REACT_APP_FF_SKINS": "1",
-      "REACT_APP_FF_RARITIES_WITH_ORACLE": "1",
       "REACT_APP_EXPLORER_URL": "https://play.decentraland.org",
       "REACT_APP_TRANSACTIONS_API_URL": "https://transactions-api.decentraland.io/v1",
       "REACT_APP_ACCOUNT_URL": "https://account.decentraland.org",
