+++ package.json
@@ -38,7 +38,7 @@
     "decentraland-ad": "^1.4.2",
     "decentraland-builder-scripts": "^0.24.0",
     "decentraland-commons": "^5.1.0",
-    "decentraland-dapps": "^12.50.0",
+    "decentraland-dapps": "^12.51.0",
     "decentraland-ecs": "^6.6.1-20201020183014.commit-bdc29ef",
     "decentraland-experiments": "^1.0.2",
     "decentraland-transactions": "^1.37.0",
