+++ src/components/Modals/PublishCollectionModal/PublishCollectionModalOld/index.ts
@@ -1,0 +1,3 @@
+import PublishCollectionModal from './PublishCollectionModal.container'
+
+export default PublishCollectionModal
