+++ src/components/Modals/index.ts
@@ -1,8 +1,3 @@
-import { isUsingRaritiesWithOracle } from 'modules/item/utils'
-
-import _PublishCollectionModalWithOracle from './PublishCollectionModalWithOracle'
-import _PublishCollectionModal from './PublishCollectionModal'
-
 export { default as AddExistingItemModal } from './AddExistingItemModal'
 export { default as ApprovalFlowModal } from './ApprovalFlowModal'
 export { default as ClaimNameFatFingerModal } from './ClaimNameFatFingerModal'
@@ -29,9 +24,7 @@ export { default as ImportModal } from './ImportModal'
 export { default as LikeModal } from './LikeModal'
 export { default as ManageCollectionRoleModal } from './ManageCollectionRoleModal'
 export { default as MintItemsModal } from './MintItemsModal'
-
-export const PublishCollectionModal = isUsingRaritiesWithOracle() ? _PublishCollectionModalWithOracle : _PublishCollectionModal
-
+export { default as PublishCollectionModal } from './PublishCollectionModal'
 export { default as PublishThirdPartyCollectionModal } from './PublishThirdPartyCollectionModal'
 export { default as PushCollectionChangesModal } from './PushCollectionChangesModal'
 export { default as QuotaExceededModal } from './QuotaExceededModal'
