+++ src/modules/common/types.ts
@@ -9,6 +9,7 @@ import { TranslationState } from 'decentraland-dapps/dist/modules/translation/re
 import { ToastState } from 'decentraland-dapps/dist/modules/toast/reducer'
 import { WalletState } from 'decentraland-dapps/dist/modules/wallet/reducer'
 import { STORAGE_LOAD } from 'decentraland-dapps/dist/modules/storage/actions'
+import { FeaturesState } from 'decentraland-dapps/dist/modules/features/reducer'
 
 import { AssetPackState } from 'modules/assetPack/reducer'
 import { AssetState } from 'modules/asset/reducer'
@@ -73,6 +74,7 @@ export type RootState = {
   wallet: WalletState
   collectionCuration: CollectionCurationState
   itemCuration: ItemCurationState
+  features: FeaturesState
 }
 
 export type RootStore = Store<RootState>
