+++ src/modules/features/types.ts
@@ -1,0 +1,3 @@
+export enum FeatureName {
+  RARITIES_WITH_ORACLE = 'rarities-with-oracle'
+}
